Aniadiendo ResponseFactory a mi rama

+++ expedienteclinico/src/main/java/com/expedienteclinico/expedienteclinico/repositories/IFirstRepository.java
@@ -1,8 +1,7 @@
 package com.expedienteclinico.expedienteclinico.repositories;
-
 import com.expedienteclinico.expedienteclinico.models.FirstModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IFirstRepository extends JpaRepository <FirstModel, Integer>  {
+public interface IFirstRepository extends JpaRepository <FirstModel, Long>  {
 
 }
